refactor(tracing): consolidate the package into four responsibility files
The package had grown to nine non-test files, several of them small and
overlapping, which made it hard to tell where a given helper belongs.
Group them by responsibility instead:

- span.go: the instrumentation API (StartReconcileSpan, StartControllerSpan,
  StartManagerRootSpan, StartManagerSpan, EndSpan) together with the Span
  name and attribute key constants it is used with — merges reconcile.go
  and spans.go
- processor.go: everything deciding what reaches the exporter — the
  FilteringSpanProcessor, the per-Reconcile write flag, and the
  write-tracking client — absorbs client.go
- provider.go: TracerProvider setup, exporters, the sampler, and the
  request-ID helpers backing the TraceID == request ID contract —
  absorbs requestid.go and sampler.go
- propagator.go: unchanged, W3C trace-context injection/extraction

Test files follow the same grouping. Update doc.go, the two tracing design
specs and one comment in framework.go to point at the new locations.

This is a pure file reorganization: no exported or unexported symbol
changed, and all 63 test and benchmark functions are preserved verbatim.
