+++ pkg/tracing/processor.go
@@ -21,6 +21,7 @@ import (
 	"sync/atomic"
 
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 // filteringSpanProcessor wraps another SpanProcessor and drops Spans that were
@@ -89,7 +90,7 @@ func withWriteFlag(ctx context.Context) context.Context {
 
 // markWrite records that a real write operation happened in the current Reconcile.
 // It is deliberately unexported: the only marking path is the write-tracking
-// client (see client.go) intercepting Kubernetes write calls, so instrumentation
+// client (see NewWriteTrackingClient) intercepting Kubernetes write calls, so instrumentation
 // authors never have to know about the flag — any write issued through the
 // client is tracked automatically.
 // It is a no-op if the context carries no write flag (e.g. tracing disabled or
@@ -128,4 +129,113 @@ func hasFailed(ctx context.Context) bool {
 		return f.failed.Load()
 	}
 	return false
+}
+
+// NewWriteTrackingClient wraps c so that every Kubernetes write issued through
+// it (Create, Update, Patch, Delete, DeleteAllOf, and subresource writes such
+// as Status().Patch) marks the current Reconcile iteration as having performed
+// real work. This is the ONLY mechanism that marks a Reconcile as a write:
+// instrumentation authors never deal with write marking — any write that goes
+// through the client is tracked automatically, and Spans (StartControllerSpan
+// + EndSpan) are purely observational.
+//
+// A write is counted when the write method is invoked, regardless of the
+// result: the request reached the API server, so the iteration did real work
+// worth retaining (a failed write additionally retains the iteration via the
+// failed flag when its error ends a Span or the Reconcile).
+//
+// When tracing is disabled the original client is returned unwrapped, so the
+// call path is structurally identical to a build without tracing — the same
+// zero-overhead philosophy as the no-op filter and sampler, which are not
+// installed at all in mode "none". This requires InitTracerProvider to have
+// run before controllers are assembled, which is its documented contract.
+//
+// Reads (Get, List, Watch) are forwarded without interception.
+func NewWriteTrackingClient(c client.Client) client.Client {
+	if !Enabled() {
+		return c
+	}
+	return &writeTrackingClient{Client: c}
+}
+
+// writeTrackingClient decorates client.Client, marking the per-Reconcile write
+// flag on every write-verb call. markWrite is a no-op when the context carries
+// no write flag (e.g. outside a Reconcile), so sharing the wrapped client with
+// non-Reconcile code paths is safe.
+type writeTrackingClient struct {
+	client.Client
+}
+
+func (c *writeTrackingClient) Create(ctx context.Context, obj client.Object, opts ...client.CreateOption) error {
+	markWrite(ctx)
+	return c.Client.Create(ctx, obj, opts...)
+}
+
+func (c *writeTrackingClient) Update(ctx context.Context, obj client.Object, opts ...client.UpdateOption) error {
+	markWrite(ctx)
+	return c.Client.Update(ctx, obj, opts...)
+}
+
+func (c *writeTrackingClient) Patch(ctx context.Context, obj client.Object, patch client.Patch, opts ...client.PatchOption) error {
+	markWrite(ctx)
+	return c.Client.Patch(ctx, obj, patch, opts...)
+}
+
+func (c *writeTrackingClient) Delete(ctx context.Context, obj client.Object, opts ...client.DeleteOption) error {
+	markWrite(ctx)
+	return c.Client.Delete(ctx, obj, opts...)
+}
+
+func (c *writeTrackingClient) DeleteAllOf(ctx context.Context, obj client.Object, opts ...client.DeleteAllOfOption) error {
+	markWrite(ctx)
+	return c.Client.DeleteAllOf(ctx, obj, opts...)
+}
+
+func (c *writeTrackingClient) Status() client.SubResourceWriter {
+	return &writeTrackingSubResourceWriter{writer: c.Client.Status()}
+}
+
+func (c *writeTrackingClient) SubResource(subResource string) client.SubResourceClient {
+	return &writeTrackingSubResourceClient{SubResourceClient: c.Client.SubResource(subResource)}
+}
+
+// writeTrackingSubResourceWriter marks writes issued via Status().
+type writeTrackingSubResourceWriter struct {
+	writer client.SubResourceWriter
+}
+
+func (w *writeTrackingSubResourceWriter) Create(ctx context.Context, obj client.Object, subResource client.Object, opts ...client.SubResourceCreateOption) error {
+	markWrite(ctx)
+	return w.writer.Create(ctx, obj, subResource, opts...)
+}
+
+func (w *writeTrackingSubResourceWriter) Update(ctx context.Context, obj client.Object, opts ...client.SubResourceUpdateOption) error {
+	markWrite(ctx)
+	return w.writer.Update(ctx, obj, opts...)
+}
+
+func (w *writeTrackingSubResourceWriter) Patch(ctx context.Context, obj client.Object, patch client.Patch, opts ...client.SubResourcePatchOption) error {
+	markWrite(ctx)
+	return w.writer.Patch(ctx, obj, patch, opts...)
+}
+
+// writeTrackingSubResourceClient marks writes issued via SubResource(...);
+// its read method (Get) is forwarded by the embedded client untouched.
+type writeTrackingSubResourceClient struct {
+	client.SubResourceClient
+}
+
+func (c *writeTrackingSubResourceClient) Create(ctx context.Context, obj client.Object, subResource client.Object, opts ...client.SubResourceCreateOption) error {
+	markWrite(ctx)
+	return c.SubResourceClient.Create(ctx, obj, subResource, opts...)
+}
+
+func (c *writeTrackingSubResourceClient) Update(ctx context.Context, obj client.Object, opts ...client.SubResourceUpdateOption) error {
+	markWrite(ctx)
+	return c.SubResourceClient.Update(ctx, obj, opts...)
+}
+
+func (c *writeTrackingSubResourceClient) Patch(ctx context.Context, obj client.Object, patch client.Patch, opts ...client.SubResourcePatchOption) error {
+	markWrite(ctx)
+	return c.SubResourceClient.Patch(ctx, obj, patch, opts...)
 }
