+++ pkg/tracing/processor_test.go
@@ -31,6 +31,7 @@ import (
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	clientfake "sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	agentsv1alpha1 "github.com/openkruise/agents/api/v1alpha1"
@@ -514,4 +515,125 @@ func TestWriteFlag_IndependentFlags(t *testing.T) {
 	markWrite(ctx1)
 	assert.True(t, hasWrite(ctx1), "ctx1 should be marked")
 	assert.False(t, hasWrite(ctx2), "ctx2 should not be affected by marking ctx1")
+}
+
+// withTracingEnabled flips the enabled flag for the duration of a test.
+func withTracingEnabled(t *testing.T, enabled bool) {
+	t.Helper()
+	prev := enabledFlag.Load()
+	enabledFlag.Store(enabled)
+	t.Cleanup(func() { enabledFlag.Store(prev) })
+}
+
+func TestNewWriteTrackingClient_DisabledReturnsOriginal(t *testing.T) {
+	withTracingEnabled(t, false)
+	base := clientfake.NewClientBuilder().Build()
+	assert.Same(t, base, NewWriteTrackingClient(base),
+		"with tracing disabled the original client must be returned unwrapped")
+}
+
+func TestNewWriteTrackingClient_EnabledWraps(t *testing.T) {
+	withTracingEnabled(t, true)
+	base := clientfake.NewClientBuilder().Build()
+	assert.NotSame(t, client.Client(base), NewWriteTrackingClient(base),
+		"with tracing enabled the client must be wrapped")
+}
+
+// TestWriteTrackingClient_Verbs verifies every write verb marks the
+// per-Reconcile write flag (regardless of the call outcome) and read verbs
+// never do.
+func TestWriteTrackingClient_Verbs(t *testing.T) {
+	withTracingEnabled(t, true)
+
+	pod := func() *corev1.Pod {
+		return &corev1.Pod{ObjectMeta: metav1.ObjectMeta{Name: "p", Namespace: "default"}}
+	}
+
+	tests := []struct {
+		name      string
+		call      func(ctx context.Context, c client.Client)
+		wantWrite bool
+	}{
+		{
+			name:      "Create marks write",
+			call:      func(ctx context.Context, c client.Client) { _ = c.Create(ctx, pod()) },
+			wantWrite: true,
+		},
+		{
+			name:      "Update marks write even on error",
+			call:      func(ctx context.Context, c client.Client) { _ = c.Update(ctx, pod()) },
+			wantWrite: true,
+		},
+		{
+			name: "Patch marks write even on error",
+			call: func(ctx context.Context, c client.Client) {
+				_ = c.Patch(ctx, pod(), client.MergeFrom(pod()))
+			},
+			wantWrite: true,
+		},
+		{
+			name:      "Delete marks write even when object is absent",
+			call:      func(ctx context.Context, c client.Client) { _ = c.Delete(ctx, pod()) },
+			wantWrite: true,
+		},
+		{
+			name:      "DeleteAllOf marks write",
+			call:      func(ctx context.Context, c client.Client) { _ = c.DeleteAllOf(ctx, pod()) },
+			wantWrite: true,
+		},
+		{
+			name: "Status Update marks write even on error",
+			call: func(ctx context.Context, c client.Client) {
+				_ = c.Status().Update(ctx, pod())
+			},
+			wantWrite: true,
+		},
+		{
+			name: "Status Patch marks write even on error",
+			call: func(ctx context.Context, c client.Client) {
+				_ = c.Status().Patch(ctx, pod(), client.MergeFrom(pod()))
+			},
+			wantWrite: true,
+		},
+		{
+			name: "SubResource Update marks write even on error",
+			call: func(ctx context.Context, c client.Client) {
+				_ = c.SubResource("status").Update(ctx, pod())
+			},
+			wantWrite: true,
+		},
+		{
+			name: "Get does not mark write",
+			call: func(ctx context.Context, c client.Client) {
+				_ = c.Get(ctx, client.ObjectKey{Namespace: "default", Name: "p"}, pod())
+			},
+			wantWrite: false,
+		},
+		{
+			name: "List does not mark write",
+			call: func(ctx context.Context, c client.Client) {
+				_ = c.List(ctx, &corev1.PodList{})
+			},
+			wantWrite: false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			cli := NewWriteTrackingClient(clientfake.NewClientBuilder().Build())
+			ctx := withWriteFlag(context.Background())
+			tt.call(ctx, cli)
+			assert.Equal(t, tt.wantWrite, hasWrite(ctx))
+		})
+	}
+}
+
+// TestWriteTrackingClient_NoFlagContext verifies writes outside a Reconcile
+// (no write flag in ctx) are safe no-ops for tracking.
+func TestWriteTrackingClient_NoFlagContext(t *testing.T) {
+	withTracingEnabled(t, true)
+	cli := NewWriteTrackingClient(clientfake.NewClientBuilder().Build())
+	ctx := context.Background()
+	pod := &corev1.Pod{ObjectMeta: metav1.ObjectMeta{Name: "p", Namespace: "default"}}
+	assert.NoError(t, cli.Create(ctx, pod))
+	assert.False(t, hasWrite(ctx))
 }
