+++ pkg/tracing/span_test.go
